+++ dflib/src/main/java/com/nhl/dflib/series/IntBaseSeries.java
@@ -1,26 +1,17 @@
 package com.nhl.dflib.series;
 
-import com.nhl.dflib.BooleanSeries;
-import com.nhl.dflib.DataFrame;
-import com.nhl.dflib.Index;
-import com.nhl.dflib.IntPredicate;
-import com.nhl.dflib.IntSeries;
-import com.nhl.dflib.Series;
-import com.nhl.dflib.SeriesGroupBy;
-import com.nhl.dflib.ValueMapper;
-import com.nhl.dflib.ValuePredicate;
-import com.nhl.dflib.ValueToRowMapper;
-import com.nhl.dflib.concat.SeriesConcat;
-import com.nhl.dflib.groupby.SeriesGrouper;
-import com.nhl.dflib.map.Mapper;
-import com.nhl.dflib.sample.Sampler;
+import com.nhl.dflib.*;
 import com.nhl.dflib.accumulator.BooleanAccumulator;
 import com.nhl.dflib.accumulator.IntAccumulator;
 import com.nhl.dflib.accumulator.ObjectAccumulator;
 import com.nhl.dflib.accumulator.UniqueIntAccumulator;
-import com.nhl.dflib.sort.IndexSorter;
+import com.nhl.dflib.concat.SeriesConcat;
+import com.nhl.dflib.groupby.SeriesGrouper;
+import com.nhl.dflib.map.Mapper;
+import com.nhl.dflib.sample.Sampler;
 import com.nhl.dflib.sort.IntComparator;
 import com.nhl.dflib.sort.IntTimSort;
+import com.nhl.dflib.sort.SeriesSorter;
 
 import java.util.Arrays;
 import java.util.Comparator;
@@ -155,7 +146,7 @@ public abstract class IntBaseSeries implements IntSeries {
     }
 
     private IntSeries doSortIndexInt(IntComparator comparator) {
-        int[] mutableIndex = IndexSorter.rowNumberSequence(size());
+        int[] mutableIndex = SeriesSorter.rowNumberSequence(size());
         IntTimSort.sort(mutableIndex, comparator);
         return new IntArraySeries(mutableIndex);
     }