+++ dflib/src/main/java/com/nhl/dflib/window/RowNumberer.java
@@ -5,7 +5,7 @@ import com.nhl.dflib.IntSeries;
 import com.nhl.dflib.concat.SeriesConcat;
 import com.nhl.dflib.series.IntSequenceSeries;
 import com.nhl.dflib.sort.IntComparator;
-import com.nhl.dflib.sort.PerColumnSorter;
+import com.nhl.dflib.sort.DataFrameSorter;
 
 import java.util.Collection;
 
@@ -40,7 +40,7 @@ public class RowNumberer {
         // (1) is good for producing a sorted Series or DataFrame, (2) is good for producing row numbers that follow
         // the sorting.. We have case (2) here.
 
-        IntSeries rowPositions = new PerColumnSorter(dataFrame).sortedPositions(sorter).sortIndexInt();
+        IntSeries rowPositions = new DataFrameSorter(dataFrame).sortedPositions(sorter).sortIndexInt();
 
         // since we control select indices, and don't expect negative values, we can safely cast to IntSeries
         return (IntSeries) RowNumberer.sequence(dataFrame.height()).select(rowPositions);