+++ RELEASE-NOTES.md
@@ -6,6 +6,7 @@
 * #120 DataFrame.pivot(..)
 * #121 BigDecimal aggregators with controlled scale
 * #123 Improve performance of "DF.filterRows(BooleanSeries)" 
+* #125 Column-based sort algorithm for better performance and expressions support
 
 ## Release 0.10